move style file to a separate folder
