+++ src/features/empires/components/Loader.js
@@ -1,4 +1,4 @@
-import styles from './Loader.module.css';
+import styles from '../css/Loader.module.css';
 
 export default function Loader() {
   return (
