+++ src/features/empires/css/EmpireDetails.module.css
